Fixed model/data directories and deletions, also stopped the deletion function from deleting .gitkeep files

+++ utils/pathutils.py
@@ -5,11 +5,12 @@ import os
 def delete_with_override(path):
     for root, dirs, files in os.walk(path, topdown=False):
         for name in files:
-            filename = os.path.join(root, name)
-            os.chmod(filename, stat.S_IWUSR)
-            os.remove(filename)
+            if not name == ".gitkeep":
+                filename = os.path.join(root, name)
+                os.chmod(filename, stat.S_IWUSR)
+                os.remove(filename)
         for name in dirs:
             os.chmod(os.path.join(root, name), stat.S_IWUSR)
             os.rmdir(os.path.join(root, name))
-    os.chmod(path, stat.S_IWUSR)
-    os.rmdir(path)
+    # os.chmod(path, stat.S_IWUSR)
+    # os.rmdir(path)
